requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.2
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
dulwich==0.21.6
fastapi==0.104.1
fastjsonschema==2.19.0
filelock==3.13.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.0
langchain==0.0.336
langsmith==0.0.64
marshmallow==3.20.1
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.2
packaging==23.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.6.1
pyngrok==7.0.3
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
slack-bolt==1.18.1
slack-sdk==3.26.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.2
tomlkit==0.12.3
tqdm==4.66.1
trove-classifiers==2023.11.14
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.24.6
yarl==1.9.2
zipp==3.17.0
